apps/ml-service/requirements.txt
fastapi
uvicorn
torch
torch-geometric
pandas
numpy
psycopg2-binary